Move GlobalPolicyIdService to be under group
